bugfix in encx24j600_rx_buf_read()

+++ firmware/os/device/encx24j600.h
@@ -26,7 +26,7 @@
 /* This macro provides the offset of register x from the start of the controller's memory map */
 #define ENCX24_SFR_OFFSET(x)	    (ENCX24_SFR_BASE + (((u32) (x)) << ENCX24_SFR_SHIFT))
 
-#define ENCX24_SFR_ADDR(base, x)    ((base) + ENCX24_SFR_OFFSET(x))
+#define ENCX24_SFR_ADDR(base, x)    (((u32) base) + ENCX24_SFR_OFFSET(x))
 
 /* This macro is an accessor for register x, given a controller at address "base" */
 #define ENCX24_REG(base, x)         *((vu16 *) ENCX24_SFR_ADDR((base), (x)))
